+++ nano_installer/bin/frontend_files/resource/js/nano.js
@@ -145,10 +145,26 @@ N.TagAddressRange = "address_range";
 N.TagAllocated = "allocated";
 N.TagGateway = "gateway";
 N.TagDNS = "dns";
+N.TagRole = "role";
+N.TagUserGroup = "user_group";
+N.TagDisplayName = "display_name";
+N.TagMember = "member";
+N.TagNickname = "nickname";
+N.TagMail = "mail";
+
+N.TagCurrent = "current";
+N.TagPrevious = "previous";
+N.TagNext = "next";
+N.TagNew = "new";
+N.TagLogin = "login";
+N.TagLogout = "logout";
+N.TagResetSystem = "reset_system";
 
 //Language name
 N.zh_CN = "zh_CN";
 N.en_US = "en_US";
+
+N._session_tag = "nano-session";
 
 var zh_CN_text = new TextSet(new Map([
   [N.TagDashboard, "仪表盘"],
@@ -259,7 +275,20 @@ var zh_CN_text = new TextSet(new Map([
   [N.TagAddressRange, "地址范围"],
   [N.TagAllocated, "已分配"],
   [N.TagGateway, "网关"],
-  [N.TagDNS, "域名服务器"]
+  [N.TagDNS, "域名服务器"],
+  [N.TagRole, "角色"],
+  [N.TagUserGroup, "用户组"],
+  [N.TagDisplayName, "显示名称"],
+  [N.TagMember, "成员"],
+  [N.TagNickname, "昵称"],
+  [N.TagMail, "邮箱"],
+  [N.TagCurrent, "当前"],
+  [N.TagPrevious, "上一个"],
+  [N.TagNext, "下一个"],
+  [N.TagNew, "新"],
+  [N.TagLogin, "登录"],
+  [N.TagLogout, "注销"],
+  [N.TagResetSystem, "重置系统"]
 ]));
 
 var en_US_text = new TextSet(new Map([
@@ -371,7 +400,20 @@ var en_US_text = new TextSet(new Map([
   [N.TagAddressRange, "Address Range"],
   [N.TagAllocated, "Allocated"],
   [N.TagGateway, "Gateway"],
-  [N.TagDNS, "DNS"]
+  [N.TagDNS, "DNS"],
+  [N.TagRole, "Role"],
+  [N.TagUserGroup, "User Group"],
+  [N.TagDisplayName, "Display Name"],
+  [N.TagMember, "Member"],
+  [N.TagNickname, "Nickname"],
+  [N.TagMail, "Mail"],
+  [N.TagCurrent, "Current "],
+  [N.TagPrevious, "Previous "],
+  [N.TagNext, "Next"],
+  [N.TagNew, "New "],
+  [N.TagLogin, "Login"],
+  [N.TagLogout, "Logout"],
+  [N.TagResetSystem, "Reset System"]
 ]));
 
 //initial text sets
@@ -380,6 +422,17 @@ N._text_sets = new Map([
   [N.zh_CN, zh_CN_text],
   [N.en_US, en_US_text]
 ]);
+
+N._menu_defines = [
+  ['dashboard', 'dashboard.html', 'multiline_chart', N.TagDashboard],
+  ['compute_pool', 'compute_pools.html', 'blur_on', N.TagResourcePool],
+  ['address_pool', 'address_pools.html', 'share', N.TagAddressPool],
+  ['storage_pool', 'storages.html', 'storage', N.TagStoragePool],
+  ['instance', 'instances.html', 'cloud', N.TagInstance],
+  ['image', 'images.html', 'content_copy', N.TagImage],
+  ['media', 'medias.html', 'album', N.TagMedia],
+  ['user', 'users.html', 'people', N.TagUser]
+];
 
 N.SetLanguage = function(lang){
   if (lang != this.zh_CN & lang != this.en_US){
@@ -418,36 +471,134 @@ N.GetTexts = function(){
   return this._text_sets.get(this.GetLanguage());
 }
 
-N.CreateMenuAndFooter = function(){
-  //menu
-  var _menuDefine = [
-    ['dashboard.html', 'multiline_chart', this.TagDashboard],
-    ['compute_pools.html', 'blur_on', this.TagResourcePool],
-    ['address_pools.html', 'share', this.TagAddressPool],
-    ['storages.html', 'storage', this.TagStoragePool],
-    ['instances.html', 'cloud', this.TagInstance],
-    ['images.html', 'content_copy', this.TagImage],
-    ['medias.html', 'album', this.TagMedia],
-    ['#', 'people', this.TagUser]
-  ];
+function _redirectToLogin(){
+  localStorage.setItem(N._session_tag, "");
+  var target = '/login.html?previous=' + encodeURIComponent(window.location.pathname);
+  window.location.replace(target);
+}
+
+function _updateSession(sessionID){
+  $.ajax({
+    url: '/sessions/'+ sessionID,
+    method: "PUT",
+    dataType: "json",
+    data: "",
+    success: function (result) {
+      if (0 != result['error_code']) {
+        M.toast({html: 'update session fail: ' + result['message']});
+        setTimeout(_redirectToLogin, 500);
+        return;
+      }
+    },
+    error: function (jqXHR, status, error) {
+      M.toast({html: 'request update session fail: ' + error, outDuration: 600});
+      setTimeout(_redirectToLogin, 500);
+    }
+  });
+}
+
+function _logoutSession(){
+  localStorage.setItem(N._session_tag, "");
+  var target = '/login.html';
+  window.location.replace(target);
+}
+
+N.SaveSession = function(id, user, menu, timeout){
+  var session = new Object();
+  session.id = id;
+  session.user = user;
+  session.menu = menu;
+  session.timeout = timeout;
+  localStorage.setItem(N._session_tag, JSON.stringify(session));
+}
+
+N.GetCurrentUser = function(){
+  var sessionString = localStorage.getItem(N._session_tag);
+  if (!sessionString || 0 == sessionString.length){
+    return null;
+  }
+  var session = JSON.parse(sessionString);
+  return session.user;
+}
+
+N.ValidateSession = function(){
+  var sessionString = localStorage.getItem(N._session_tag);
+  if (!sessionString || 0 == sessionString.length){
+    //no session available
+    _redirectToLogin();
+    return
+  }
+  var session = JSON.parse(sessionString);
+  if (!session.id){
+    _redirectToLogin();
+    return
+  }
+  //try update
+  var updateInteval = session.timeout / 2;
+  $.ajax({
+    url: '/sessions/'+ session.id,
+    method: "PUT",
+    dataType: "json",
+    data: "",
+    success: function (result) {
+      if (0 != result['error_code']) {
+        M.toast({html: 'update session fail: ' + result['message']});
+        setTimeout(_redirectToLogin, 500);
+        return;
+      }
+      //build menu
+      N.CreateMenuAndFooter(session.user, session.menu);
+      var update = function(){
+        _updateSession(session.id);
+      };
+      //keep updating
+      setInterval(update, updateInteval * 1000);
+    },
+    error: function (jqXHR, status, error) {
+      M.toast({html: 'request update session fail: ' + error, outDuration: 600});
+      setTimeout(_redirectToLogin, 500);
+    }
+  });
+
+}
+
+N.GetAllMenus = function(){
+  var result = new Array();
+  this._menu_defines.forEach((menu)=>{
+    var name = menu[0];
+    var tag = menu[3];
+    result.push([name, this.GetTexts().get(tag)]);
+  });
+  return result;
+}
+
+N.CreateMenuAndFooter = function(userName, menuList){
   var texts = this.GetTexts();
   var menu = $('<ul>').addClass('right');
-  _menuDefine.forEach((item) =>{
-    menu.append(
-      $('<li>').addClass('hide-on-small-only').append(
-        $('<a>').attr('href', item[0]).text(texts.get(item[2])).prepend(
-          $('<i>').addClass('material-icons left').text(item[1])
+
+  if (menuList){
+    var selectedMenu = new Set();
+    menuList.forEach((menuName)=>{
+      selectedMenu.add(menuName);
+    });
+    this._menu_defines.forEach((define)=>{
+      var menuName = define[0];
+      if (selectedMenu.has(menuName)){
+        var page = define[1];
+        var icon = define[2];
+        var label = texts.get(define[3]);
+        menu.append(
+          $('<li>').addClass('hide-on-small-only').append(
+            $('<a>').attr('href', page).text(label).prepend(
+              $('<i>').addClass('material-icons left').text(icon)
+            )
+          )
         )
-      )
-    )
-  });
-  menu.append(
-    $('<li>').addClass('hide-on-med-and-up').append(
-      $('<a>').attr('href', '#').prepend(
-        $('<i>').addClass('material-icons right').text('menu')
-      ).text(texts.get(this.TagMenu))
-    )
-  );
+      }
+    });
+  }
+
+
   $('body').prepend(
     $('<nav>').addClass('blue-grey').append(
       $('<div>').addClass('container').append(
@@ -477,17 +628,33 @@ N.CreateMenuAndFooter = function(){
   };
   langSwitch.click(onLanguageChanged);
 
+  var footerContainer = $('<div>').addClass('container').text('© 2018 Copyright Project Nano. All rights reserved.');
+  var rightDiv = $('<div>').addClass('row');
+  if (userName){
+    rightDiv.append(
+      $('<div>').addClass('col m2 push-m8').append(
+        $('<label>').addClass('white-text').text(texts.get(this.TagCurrent) + texts.get(this.TagUser) + ': ' + userName)
+      ).append(
+        $('<a>').addClass('btn-small btn-flat').attr('onclick', '_logoutSession()').append(
+          $('<i>').addClass('material-icons white-text').text('exit_to_app')
+        )
+      )
+    );
+  }
+  rightDiv.append(
+    $('<div>').addClass('switch col m2 push-m8').append(
+      $('<label>').addClass('white-text').append('English').append(
+        langSwitch
+      ).append(
+        $('<span>').addClass('lever')
+      ).append('中文')
+    )
+  );
   $('body').append(
     $('<footer>').addClass('page-footer blue-grey darken-3').append(
       $('<div>').addClass('footer-copyright').append(
-        $('<div>').addClass('container').text('© 2018 Copyright Project Nano. All rights reserved.').append(
-          $('<div>').addClass('switch right').append(
-            $('<label>').addClass('white-text').append('English').append(
-              langSwitch
-            ).append(
-              $('<span>').addClass('lever')
-            ).append('中文')
-          )
+        footerContainer.append(
+          rightDiv
         )
       )
     )